requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
colorama==0.4.6
cssselect2==0.8.0
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fonttools==4.58.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipdb==0.13.13
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
openai==1.79.0
pango==0.0.1
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
pypandoc==1.15
pyphen==0.17.2
python-docx==1.1.2
PyYAML==6.0
referencing==0.36.2
reportlab==4.4.1
resumy==0.1.2
rpds-py==0.25.0
sniffio==1.3.1
stack-data==0.6.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
